requirements.txt
pandas
matplotlib
sentence-transformers
faiss-cpu
requests
python-dotenv
streamlit